chore: 바텀 모달 타입 분리
